refactor: move card creation handling logic to resolver
